Add files via upload
I am learning how to work with Python still.  I wanted to show an example of branching within a simple number guesser.  I hope to clean up my code more in the future.  If you're wondering why my work is all in Camel Case, it's because my professor wants us to pick up the habit for when we work with Java.  Play nice with the program and (generally) enter what it asks.  I haven't figured out how to stop ValueErrors yet, especially within a loop.  I hope to change that very soon.  Until then, if you play with this code - please type in the specified values.  Although there is a catch for numbers below 1 and higher than 20.